# src/web/openapi.yaml
openapi: 3.0.3
info:
  title: Attesting API
  description: OSCAL-native GRC platform — governance, risk, compliance, threat intelligence, and drift detection.
  version: 0.3.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development
paths:
...
  /import/preview:
    post:
      summary: Upload and preview catalog import
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Preview with controls and mappings
        "400":
          description: File rejected by security scan
  /import/confirm:
    post:
      summary: Execute a previewed import
      tags: [Import]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upload_path, original_name]
              properties:
                upload_path: { type: string }
                original_name: { type: string }
                format: { type: string }
                overwrite: { type: boolean }
      responses:
        "201":
          description: Import result
...